Triangle deduplication for colorization
Triangles shapes can be described as the differences between the
vertex positions in a 1D representation of the image. These
differences are combined into an integer. Then, unique elements
can be found, which describe base triangles. For those, the
positions inside the triangle can be calculated. Duplicate
triangles then share one and the same calculation result.